requirements.txt
mkdocs-material
mkdocs-gen-files
mkdocs-macros-plugin
mkdocs-simple-hooks
neoteroi-mkdocs
requests
PyYAML